Backend/requirements.txt
fastapi
uvicorn
motor
python-jose
passlib[bcrypt]
python-dotenv
pydantic[email]
pydantic-settings